patch up after rename
