+++ esmodules/main.js
@@ -9,8 +9,11 @@ import {
   isPerceptionActive,
   clearPf2ePerceptionFlags,
 } from "./pf2e_perception.js";
-import { registerHooksForClearMovementHistory } from "./clearMovement.js";
-import { clearPartyStealth, clearTokenStealth } from "./clearStealth.js";
+import { registerHooksForClearMovementHistory } from "./clear-movement.js/index.js";
+import {
+  clearPartyStealth,
+  clearTokenStealth,
+} from "./clear-stealth.js/index.js";
 
 function colorizeOutput(format, ...args) {
   return [`%c${MODULE_ID} %c|`, ...CONSOLE_COLORS, format, ...args];
